✨ Intégration éditeur visuel dans config + nettoyage pages buggées
